requirements/base.txt
#############################
#
#  edX Fork of Segment's python-analytics 1.2.8 using gevent for queueing
#
#############################
-e git+https://github.com/edx/analytics-python.git@1.2.11#egg=analytics-python==1.2.11

coreapi==2.3.1
django==1.11.15
django-compressor==2.2
django-crispy-forms==1.7.0
django_extensions==1.9.9
django-filter==1.0.4
django-libsass==0.5
django-oscar==1.5.1
django-rest-swagger==2.2.0
django-solo==1.1.2
django-threadlocals==0.8
django-waffle==0.12.0
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
drf-extensions==0.3.1
edx-auth-backends==1.1.3
edx-django-release-util==0.3.1
edx-django-sites-extensions==1.0.0
edx-drf-extensions==1.2.3
edx-ecommerce-worker==0.6.0
edx-opaque-keys==0.3.1
edx-rest-api-client==1.7.1
jsonfield==1.0.3
libsass==0.9.2
markdown==2.6.9
ndg-httpsclient==0.4.2
path.py==7.2
paypalrestsdk==1.11.5
premailer==2.9.2
pycountry==17.1.8
pygments==2.2.0
python-dateutil==2.4.2
pytz==2016.10
requests==2.18.1
sailthru-client==2.2.3
six==1.10.0
stripe==1.70.0
zeep==2.1.1

# alipay
alipay-sdk-python==3.2.53
rsa==4.0
pycrypto==2.6.1
